requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
decorator==4.4.2
idna==2.10
joblib==0.16.0
lxml==4.5.2
nltk==3.5
pycparser==2.20
PySnooper==0.4.2
python-telegram-bot==12.8
regex==2020.7.14
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
tqdm==4.49.0
urllib3==1.25.10
